feat: added employees to the redux store
