+++ src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './ui';
 export * from './data';
+export * from './slices';
