+++ src/interfaces/slices/index.ts
@@ -1,0 +1,5 @@
+import { IEmployee } from '../data';
+
+export interface IEmployeeState {
+  employees: IEmployee[];
+}
